Introduce exec_env.py to maintain global state + minor bug fix
1. The global database_instances is moved to exec_env.py to avoid
   circular imports.

2. SpannerFiledInfo.typename populated with the correct name now

3. Remove all cloud spanner refs from database.py
